feat: Move self to personal
