+++ internal/personal/uuid.go
@@ -1,4 +1,4 @@
-package middle
+package personal
 
 import (
 	"encoding/json"
